+++ src/index.ts
@@ -2,7 +2,9 @@ export * from "./build";
 export * from "./nitro";
 export * from "./scan";
 export * from "./dev/server";
-export * from "./options";
 export * from "./types";
 export * from "./prerender";
 export * from "./preset";
+
+export { loadOptions } from "./options";
+export type { LoadConfigOptions } from "./options";