+++ src/presets/netlify.ts
@@ -2,6 +2,7 @@ import { existsSync, promises as fsp } from "node:fs";
 import { join, dirname } from "pathe";
 import { defineNitroPreset } from "../preset";
 import type { Nitro } from "../types";
+import { name, version } from "../../package.json";
 
 // Netlify functions
 export const netlify = defineNitroPreset({
@@ -17,6 +18,9 @@ export const netlify = defineNitroPreset({
     },
   },
   hooks: {
+    "rollup:before": (nitro: Nitro) => {
+      deprecateSWR(nitro);
+    },
     async compiled(nitro: Nitro) {
       await writeHeaders(nitro);
       await writeRedirects(nitro);
@@ -38,6 +42,11 @@ export const netlify = defineNitroPreset({
 export const netlifyBuilder = defineNitroPreset({
   extends: "netlify",
   entry: "#internal/nitro/entries/netlify-builder",
+  hooks: {
+    "rollup:before": (nitro: Nitro) => {
+      deprecateSWR(nitro);
+    },
+  },
 });
 
 // Netlify edge
@@ -45,23 +54,32 @@ export const netlifyEdge = defineNitroPreset({
   extends: "base-worker",
   entry: "#internal/nitro/entries/netlify-edge",
   output: {
-    serverDir: "{{ rootDir }}/.netlify/edge-functions",
+    serverDir: "{{ rootDir }}/.netlify/edge-functions/server",
     publicDir: "{{ rootDir }}/dist",
   },
   rollupConfig: {
     output: {
-      entryFileNames: "server.mjs",
+      entryFileNames: "server.js",
       format: "esm",
     },
   },
+  unenv: {
+    polyfill: ["#internal/nitro/polyfill/deno-env"],
+  },
   hooks: {
+    "rollup:before": (nitro: Nitro) => {
+      deprecateSWR(nitro);
+    },
     async compiled(nitro: Nitro) {
+      // https://docs.netlify.com/edge-functions/create-integration/
       const manifest = {
         version: 1,
         functions: [
           {
+            path: "/*",
+            name: "nitro server handler",
             function: "server",
-            pattern: "^.*$",
+            generator: `${name}@${version}`,
           },
         ],
       };
@@ -84,6 +102,9 @@ export const netlifyStatic = defineNitroPreset({
     preview: "npx serve ./static",
   },
   hooks: {
+    "rollup:before": (nitro: Nitro) => {
+      deprecateSWR(nitro);
+    },
     async compiled(nitro: Nitro) {
       await writeHeaders(nitro);
       await writeRedirects(nitro);
@@ -98,9 +119,9 @@ async function writeRedirects(nitro: Nitro) {
   )
     ? "/* /404.html 404"
     : "";
-  let contents = !nitro.options.static
-    ? "/* /.netlify/functions/server 200"
-    : staticFallback;
+  let contents = nitro.options.static
+    ? staticFallback
+    : "/* /.netlify/functions/server 200";
 
   const rules = Object.entries(nitro.options.routeRules).sort(
     (a, b) => a[0].split(/\/(?!\*)/).length - b[0].split(/\/(?!\*)/).length
@@ -183,4 +204,35 @@ async function writeHeaders(nitro: Nitro) {
   }
 
   await fsp.writeFile(headersPath, contents);
+}
+
+function deprecateSWR(nitro: Nitro) {
+  if (nitro.options.future.nativeSWR) {
+    return;
+  }
+  let hasLegacyOptions = false;
+  for (const [key, value] of Object.entries(nitro.options.routeRules)) {
+    if (_hasProp(value, "isr")) {
+      continue;
+    }
+    if (value.cache === false) {
+      value.isr = false;
+    }
+    if (_hasProp(value, "static")) {
+      value.isr = !(value as { static: boolean }).static;
+    }
+    if (value && value.cache && _hasProp(value.cache, "swr")) {
+      value.isr = value.cache.swr;
+    }
+    hasLegacyOptions = hasLegacyOptions || _hasProp(value, "isr");
+  }
+  if (hasLegacyOptions) {
+    console.warn(
+      "[nitro] Nitro now uses `isr` option to configure ISR behavior on Netlify. Backwards-compatible support for `static` and `swr` support with Builder Functions will be removed in the future versions. Set `future.nativeSWR: true` nitro config disable this warning."
+    );
+  }
+}
+
+function _hasProp(obj: any, prop: string) {
+  return obj && typeof obj === "object" && prop in obj;
 }