+++ src/types/index.ts
@@ -1,6 +1,6 @@
 export * from "./fetch";
+export * from "./global";
 export * from "./serialize";
 export * from "./nitro";
 export * from "./handler";
 export * from "./utils";
-export * from "../runtime/types";