+++ src/types/nitro.ts
@@ -11,6 +11,8 @@ import type { RollupWasmOptions } from "@rollup/plugin-wasm";
 import type { Storage, BuiltinDriverName } from "unstorage";
 import type { ServerOptions as HTTPProxyOptions } from "http-proxy";
 import type { ProxyOptions } from "h3";
+import type { ResolvedConfig, ConfigWatcher } from "c12";
+import type { TSConfig } from "pkg-types";
 import type { NodeExternalsOptions } from "../rollup/plugins/externals";
 import type { RollupConfig } from "../rollup/config";
 import type { Options as EsbuildOptions } from "../rollup/plugins/esbuild";
@@ -24,6 +26,27 @@ import type {
 import type { PresetOptions } from "./presets";
 import type { KebabCase } from "./utils";
 
+export type NitroDynamicConfig = Pick<
+  NitroConfig,
+  "runtimeConfig" | "routeRules"
+>;
+
+export interface NitroRuntimeConfigApp {
+  baseURL: string;
+  [key: string]: any;
+}
+
+export interface NitroRuntimeConfig {
+  app: NitroRuntimeConfigApp;
+  nitro: {
+    envPrefix?: string;
+    routeRules?: {
+      [path: string]: NitroRouteConfig;
+    };
+  };
+  [key: string]: any;
+}
+
 export interface Nitro {
   options: NitroOptions;
   scannedHandlers: NitroEventHandler[];
@@ -33,6 +56,7 @@ export interface Nitro {
   logger: ConsolaInstance;
   storage: Storage;
   close: () => Promise<void>;
+  updateConfig: (config: NitroDynamicConfig) => void | Promise<void>;
 
   /* @internal */
   _prerenderedRoutes?: PrerenderGenerateRoute[];
@@ -53,9 +77,11 @@ export interface PrerenderGenerateRoute extends PrerenderRoute {
 
 type HookResult = void | Promise<void>;
 export interface NitroHooks {
-  "rollup:before": (nitro: Nitro) => HookResult;
+  "rollup:before": (nitro: Nitro, config: RollupConfig) => HookResult;
   compiled: (nitro: Nitro) => HookResult;
   "dev:reload": () => HookResult;
+  "rollup:reload": () => HookResult;
+  restart: () => HookResult;
   close: () => HookResult;
   "prerender:routes": (routes: Set<string>) => HookResult;
   "prerender:route": (route: PrerenderRoute) => HookResult;
@@ -153,6 +179,7 @@ export interface NitroRouteRules
 export interface NitroOptions extends PresetOptions {
   // Internal
   _config: NitroConfig;
+  _c12: ResolvedConfig<NitroConfig> | ConfigWatcher<NitroConfig>;
 
   // General
   debug: boolean;
@@ -160,12 +187,7 @@ export interface NitroOptions extends PresetOptions {
   preset: KebabCase<keyof typeof _PRESETS> | (string & {});
   static: boolean;
   logLevel: LogLevel;
-  runtimeConfig: {
-    app: {
-      baseURL: string;
-    };
-    [key: string]: any;
-  };
+  runtimeConfig: NitroRuntimeConfig;
   appConfig: AppConfig;
   appConfigFiles: string[];
 
@@ -192,6 +214,10 @@ export interface NitroOptions extends PresetOptions {
   experimental?: {
     wasm?: boolean | RollupWasmOptions;
     legacyExternals?: boolean;
+    openAPI?: boolean;
+  };
+  future: {
+    nativeSWR: boolean;
   };
   serverAssets: ServerAssetDir[];
   publicAssets: PublicAssetDir[];
@@ -215,7 +241,10 @@ export interface NitroOptions extends PresetOptions {
   errorHandler: string;
   devErrorHandler: NitroErrorHandler;
   prerender: {
+    concurrency: number;
+    interval: number;
     crawlLinks: boolean;
+    failOnError: boolean;
     ignore: string[];
     routes: string[];
   };
@@ -246,6 +275,7 @@ export interface NitroOptions extends PresetOptions {
     generateTsConfig?: boolean;
     /** the path of the generated `tsconfig.json`, relative to buildDir */
     tsconfigPath: string;
+    tsConfig?: Partial<TSConfig>;
   };
   hooks: NestedHooks<NitroHooks>;
   nodeModulesDirs: string[];
@@ -253,4 +283,8 @@ export interface NitroOptions extends PresetOptions {
     preview: string;
     deploy: string;
   };
+}
+
+declare global {
+  const defineNitroConfig: (config: NitroConfig) => NitroConfig;
 }