+++ test/tests.ts
@@ -2,7 +2,7 @@ import { resolve } from "pathe";
 import { listen, Listener } from "listhen";
 import destr from "destr";
 import { fetch, FetchOptions } from "ofetch";
-import { expect, it, afterAll } from "vitest";
+import { expect, it, afterAll, beforeAll, describe } from "vitest";
 import { fileURLToPath } from "mlly";
 import { joinURL } from "ufo";
 import * as _nitro from "../src";
@@ -20,6 +20,14 @@ export interface Context {
   server?: Listener;
   isDev: boolean;
 }
+
+// https://github.com/unjs/nitro/pull/1240
+export const describeIf = (condition, title, factory) =>
+  condition
+    ? describe(title, factory)
+    : describe(title, () => {
+        it.skip("skipped", () => {});
+      });
 
 export async function setupTest(preset: string) {
   const fixtureDir = fileURLToPath(new URL("fixture", import.meta.url).href);
@@ -121,9 +129,9 @@ export function testNitro(
     };
   }
 
-  it("setup handler", async () => {
+  beforeAll(async () => {
     _handler = await getHandler();
-  });
+  }, 25_000);
 
   it("API Works", async () => {
     const { data: helloData } = await callHandler({ url: "/api/hello" });
@@ -154,6 +162,11 @@ export function testNitro(
     const obj = await callHandler({ url: "/rules/redirect/obj" });
     expect(obj.status).toBe(308);
     expect(obj.headers.location).toBe("https://nitro.unjs.io/");
+  });
+
+  it("render JSX", async () => {
+    const { data } = await callHandler({ url: "/jsx" });
+    expect(data).toMatch("<h1 >Hello JSX!</h1>");
   });
 
   it("handles route rules - headers", async () => {
@@ -249,15 +262,18 @@ export function testNitro(
       `);
     });
 
-    it("resolve module version conflicts", async () => {
-      const { data } = await callHandler({ url: "/modules" });
-      expect(data).toMatchObject({
-        depA: "nitro-lib@1.0.0+nested-lib@1.0.0",
-        depB: "nitro-lib@2.0.1+nested-lib@2.0.1",
-        depLib: "nitro-lib@2.0.0+nested-lib@2.0.0",
-        subpathLib: "nitro-lib@2.0.0",
-      });
-    });
+    it.skipIf(ctx.preset === "deno-server")(
+      "resolve module version conflicts",
+      async () => {
+        const { data } = await callHandler({ url: "/modules" });
+        expect(data).toMatchObject({
+          depA: "nitro-lib@1.0.0+nested-lib@1.0.0",
+          depB: "nitro-lib@2.0.1+nested-lib@2.0.1",
+          depLib: "nitro-lib@2.0.0+nested-lib@2.0.0",
+          subpathLib: "nitro-lib@2.0.0",
+        });
+      }
+    );
 
     it("useStorage (with base)", async () => {
       const putRes = await callHandler({
@@ -297,19 +313,53 @@ export function testNitro(
     }
   }
 
-  it("app config", async () => {
+  it("runtime proxy", async () => {
     const { data } = await callHandler({
-      url: "/app-config",
+      url: "/api/proxy?foo=bar",
+      headers: {
+        "x-test": "foobar",
+      },
     });
-    expect(data).toMatchInlineSnapshot(`
-      {
-        "appConfig": {
-          "app-config": true,
-          "nitro-config": true,
-          "server-config": true,
+    expect(data.headers["x-test"]).toBe("foobar");
+    expect(data.url).toBe("/api/echo?foo=bar");
+  });
+
+  it("config", async () => {
+    const { data } = await callHandler({
+      url: "/config",
+    });
+    expect(data).toMatchObject({
+      appConfig: {
+        dynamic: "from-middleware",
+        "app-config": true,
+        "nitro-config": true,
+        "server-config": true,
+      },
+      runtimeConfig: {
+        dynamic: "from-env",
+        app: {
+          baseURL: "/",
         },
-      }
-    `);
+      },
+      sharedAppConfig: {
+        dynamic: "initial",
+        "app-config": true,
+        "nitro-config": true,
+        "server-config": true,
+      },
+      sharedRuntimeConfig: {
+        dynamic:
+          // TODO
+          ctx.preset.includes("cloudflare") ||
+          ctx.preset === "vercel-edge" ||
+          ctx.preset === "nitro-dev"
+            ? "initial"
+            : "from-env",
+        app: {
+          baseURL: "/",
+        },
+      },
+    });
   });
 
   if (ctx.nitro!.options.timing) {
@@ -321,4 +371,21 @@ export function testNitro(
       expect(headers["server-timing"]).toMatch(/-;dur=\d+;desc="Generate"/);
     });
   }
+
+  it("static build flags", async () => {
+    const { data } = await callHandler({ url: "/static-flags" });
+    expect(data).toMatchObject({
+      dev: [ctx.isDev, ctx.isDev],
+      preset: [ctx.preset, ctx.preset],
+      prerender: [
+        ctx.preset === "nitro-prerenderer",
+        ctx.preset === "nitro-prerenderer",
+      ],
+      client: [false, false],
+      nitro: [true, true],
+      server: [true, true],
+      "versions.nitro": [expect.any(String), expect.any(String)],
+      "versions?.nitro": [expect.any(String), expect.any(String)],
+    });
+  });
 }